projekty/pwir/requirements.txt
paas-utils==0.1.1
tornado_psql_helper_tools==0.4.1
momoko==1.1.5
tornado_yieldperiodic==0.1
pyyaml
